Add infrastructure for pluggable layer-selection.
We use a Strategy pattern to separate the various layer
selection algorithms from the generic layer selection
class.

The patch already introduces a new command-line option,
"--layer-selector" and the associated response-file
syntactic comment "layer-selector" to control the
default layer selection algorithm.

THIS REVISION IS KNOWN TO BE BUGGY: Response files do
not pass the layer selector.