Fix the computation of the maximum number of pyramid levels for the umpteenth time.
The algorithm now is much more aggressive to find the highest number
of levels.  Moreover, the size of the overlap region bounding box is
increased to fix a problem with large-scale brightness differences in
some images.

This changeset could improve the quality of panoramas for all those
users who have not used the workaround "--levels=29".