Hint whether enblend(1) and enfuse(1) have been built with PREFER_SEPARATE_OPENCL_SOURCE.
Only show the OpenCL source file search path in the output of
`--version --verbose' if Enblend and Enfuse were compile with
PREFER_SEPARATE_OPENCL_SOURCE.  Otherwise the OpenCL source is inside
the binaries and the search path unneeded.  That way we can decide at
a glance whether hacking OpenCL sources affects Enblend/Enfuse without
a recompilation of the C++ sources.