Fix wrong usage of `std::unique_ptr' vs `new []' in message handler modules.
The correct template parameter for `std::unique_ptr' is `char []' if
the object is allocated with operator `new []'.  In our case it is
easier to resize the message and overwrite its contents using a
pointer from `std::string::operator[]'.  The latter is valid use of
`operator[]' and guaranteed to work by C++11.

Reviewed by Chris.