Introduce L*a*b* and L*u*v* color spaces for pyramidal blending operations in addition to the existing ones RGB and CIECAM.
In both new spaces the color difference channels a* and b* are
unlimited.  Therefore, we must limit them before conversion to the
fixed-point representation inside the pyramid.  The parameter
        lab-color-limit
controls the absolute value of the limit (default: 127.0).

In turn rework the CIECAM code to get closer to the implementations of
the Lab and Luv functors.

For the first time ever document the log-rescaling for floating-point
image data.