Beautify "numerictraits.diff" and harden it with `static_assert()'.
Introduce two alternatives to construct pyramids for 32-bit integral
images
        PREFER_INT64_TO_DOUBLE_AS_SKIPSM_IMAGE_TYPE
and double precision images
        PREFER_LONG_DOUBLE_TO_DOUBLE_AS_SKIPSM_IMAGE_TYPE
If not defined we reproduce the old behavior.

Fix a concealed glitch in the use of the template traits that relate
the image data types with the pyramid data types.

With the bug squashed we can use 64-bit *integral* arithmetic for
32-bit integer images and thus unburden the FPU a bit.  Compile with
        PREFER_DOUBLE_TO_INT64_AS_SKIPSM_IMAGE_TYPE
to get back the old data type.

Along the same lines we can now  use `long double' instead of
`double' for 64-bit floating point images.  Here we stick with the old
data type, but when compiling with
        PREFER_LONG_DOUBLE_TO_DOUBLE_AS_SKIPSM_IMAGE_TYPE
the pyramids will be built with extended precision.