Discriminate seriousness of warnings for different ICC profiles.
If we blend inside the RGB-cube, mixing profiles or alternating
between profile and no-profile images deserves a warning.  If we blend
in a dedicated colorspace (Luv, Lab, ...) all the differences will
have been transformed out for us before the blending step and any
discrepancies of profiles just deserves an `info'.