Supply dynamic TeX preamble with a "late" part.
We build some TeX code using the Make variable `DYNAMIC_TEX_PREAMBLE'
and a default part defined in the Makefile.  The code is written to
"dynamic-preamble.tex", which is included as early as possible in the
main documents "enblend.tex" and "enfuse.tex" to allow for controlling
even the document class.

However, some commands are only effective if issued _after_ document
class and packages have been loaded.  Therefore, we add the
parameter-less macro `\finishdynamicpreamble' to the default
definition of "dynamic-preamble.tex".  `\finishdynamicpreamble' is
expanded right before `\begin{document}'.