Make thresholding the alpha channels of the input images more flexible. This fixes a problem with signed alpha channels (of 16-bit images) as reported by Justin Hyland.
New parameters:
    * `import-alpha-lower-threshold', type floating-point, default: alpha maximum value / 2;
    * `import-alpha-upper-threshold', type floating-point, default: alpha maximum value;
    * `import-alpha-save-threshed', type boolean, default: false;

Enblend and Enfuse run a threshold filter on each of their input
images' alpha channels.  Its default maps alpha values from the
minimum to the half of the maximum to zero and everything above half
of the maximum to the maximum.  Parameters
`import-alpha-lower-threshold' and `import-alpha-upper-threshold'
override the defaults.

Parameter `import-alpha-save-threshed' controls whether the masks are
saved _after_ the threshold filter as "threshed-import-alpha-%d.tif".
This allows the check the filter results directly.